+++ src/components/Avatar/constants.ts
@@ -1,0 +1,13 @@
+const avatarColors = [
+  'green',
+  'purple',
+  'pink',
+  'red',
+  'orange',
+  'yellow',
+  'teal',
+  'cyan',
+  'blue',
+  'indigo',
+];
+export { avatarColors };