+++ src/components/Icon/Icon.tsx
@@ -1,50 +1,19 @@
 import React, { forwardRef } from 'react';
 
-import { useGetTheme, useTShirtSize } from '../../hooks';
 import { IconProps } from '../../types';
-import Box from '../Box';
+
+import IconDefault from './IconDefault';
+import IconSymbol from './IconSymbol';
 
 const Icon = forwardRef<HTMLElement, IconProps>((props, ref) => {
-  const theme = useGetTheme();
+  const { icon } = props;
 
-  const {
-    color,
-    icon: IconComponent,
-    sx,
-    size = theme.defaultIconSize,
-    variant,
-    title,
-    ...others
-  } = props;
+  if (typeof icon === 'string') {
+    return (<IconSymbol {...props} ref={ref} />);
+  }
 
-  const { sizeProps } = useTShirtSize({ size, sizes: theme.tShirtSizes });
-
-  const { defaultIconColor } = theme;
-
-  const resolvedTitle = title ?? (
-    typeof IconComponent === 'object' && 'type' in IconComponent
-      ? { name: IconComponent.type.name }
-      : ''
-  );
-
-  return (
-    <Box
-      as={IconComponent}
-      ref={ref}
-      role="img"
-      title={resolvedTitle}
-      variant={variant}
-      size={sizeProps.size}
-      sx={{
-        fill: color || defaultIconColor,
-        minWidth: sizeProps.size,
-        ...sx,
-      }}
-      {...others}
-    />
-  );
+  return <IconDefault {...props} ref={ref} />;
 });
 
-Icon.displayName = 'Icon';
 
 export default Icon;