+++ src/components/NavBar/stories/NavBarNextGenComponent.tsx
@@ -12,6 +12,7 @@ import { useGetTheme } from '../../../hooks';
 
 export const NavBarNextGenComponent = () => {
   const { icons } = useGetTheme();
+
   const firstSectionData = [
     {
       'data-id': 'dashboard-data-id',
@@ -84,6 +85,7 @@ export const NavBarNextGenComponent = () => {
       'data-id': 'Authentication-data-id',
       heading: 'Authentication',
       icon: icons.authenticationIcon,
+      title: 'Policies',
       key: 'Authentication',
       children: [
         <NavBarItemButton
@@ -92,6 +94,11 @@ export const NavBarNextGenComponent = () => {
         >
           Authentication Policies
         </NavBarItemButton>,
+        {
+          hasSeparator: false,
+          subTitle: 'Applications',
+        },
+        <Separator key="separator" variant="separator.navBarSubtitleSeparator" />,
         <NavBarItemButton
           key="Password-Policies"
           id="Password-Policies"