+++ CHANGELOG.md
@@ -1,3 +1,38 @@
+## 2.172.0-alpha.0 (2026-02-18)
+
+### 🚀 Features
+
+- [UIP-8012] Update Onyx theme to include Material Symbols icons
+
+### 🩹 Fixes
+
+- [UIP-8111] Astro build failure
+
+### ❤️  Thank You
+
+- Ajith M
+- Ridhisha Dangol
+
+## 2.171.0-alpha.0 (2026-02-16)
+
+### 🚀 Features
+
+- [UIP-8078] Avatar fix colors
+
+### ❤️  Thank You
+
+- Andrew Keohane
+
+## 2.170.1-alpha.0 (2026-02-16)
+
+### 🩹 Fixes
+
+- [UIP-8082] Onyx Side nav design improvements
+
+### ❤️  Thank You
+
+- Andrew Keohane
+
 ## 2.170.0 (2026-02-12)
 
 This was a version bump only for astro to align it with other projects, there were no code changes.